requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=10.0.0
PyYAML>=6.0
tqdm>=4.65.0
onnx>=1.15.0
onnxruntime>=1.17.0
huggingface-hub>=0.23.0
